// src/types/LaravelDatatableType.d.ts
export type LaravelDatatableType = {
    current_page?: number,
    data: object[],
    first_page_url?: string,
    from?: number,
    last_page?: number,
    last_page_url?: string,
    links?: object[],
    next_page_url?: string,
    path?: string,
    per_page?: number,
    prev_page_url?: string,
    to?: number,
    total?: number
}